+++ backend/src/modules/auth/services/mfa.service.ts
@@ -17,6 +17,9 @@ import {
   MfaDeviceType,
   MfaDeviceStatus,
 } from '../entities/mfa-device.entity';
+import { AuthResponseDto } from '../dto/auth-response.dto';
+import { JwtService } from '@nestjs/jwt';
+import { AuthService } from '../auth.service';
 
 @Injectable()
 export class MfaService {
@@ -29,6 +32,7 @@ export class MfaService {
     @InjectRepository(User)
     private userRepository: Repository<User>,
     private configService: ConfigService,
+    private jwtService: JwtService,
   ) {
     const key = this.configService.get<string>('SECURITY_ENCRYPTION_KEY');
     if (!key) {
@@ -40,7 +44,8 @@ export class MfaService {
   /**
    * Generate TOTP secret and QR code for a user
    */
-  async generateMfaSecret(
+  // eslint-disable-next-line prettier/prettier
+  async generateMfaSecret (
     userId: string,
     deviceName?: string,
   ): Promise<{
@@ -289,5 +294,100 @@ export class MfaService {
     let decrypted = decipher.update(encrypted, 'hex', 'utf8');
     decrypted += decipher.final('utf8');
     return decrypted;
+  }
+
+  async getMfaDevice(
+    user: User,
+    authService: AuthService,
+  ): Promise<
+    (AuthResponseDto & { mfaRequired: true; mfaToken: string }) | void
+  > {
+    // Check if MFA is enabled
+    const mfaDevice = await this.mfaDeviceRepository.findOne({
+      where: {
+        userId: user.id,
+        type: MfaDeviceType.TOTP,
+        status: MfaDeviceStatus.ACTIVE,
+      },
+    });
+
+    if (mfaDevice) {
+      // Generate temporary token for MFA verification
+      const tempToken = this.jwtService.sign(
+        {
+          sub: user,
+          email: user.email,
+          role: user.role,
+          type: 'mfa_required',
+        },
+        {
+          secret: authService.getJwtSecret(),
+          expiresIn: '5m', // Short-lived token for MFA verification
+        },
+      );
+
+      this.logger.log(`MFA required for user: ${user.id}`);
+      return {
+        user: authService.sanitizeUser(user),
+        accessToken: null,
+        refreshToken: null,
+        mfaRequired: true,
+        mfaToken: tempToken,
+      };
+    }
+  }
+
+  /**
+   * Complete login after MFA verification
+   */
+  async completeMfaLogin(
+    mfaToken: string,
+    authService: AuthService,
+  ): Promise<AuthResponseDto> {
+    try {
+      // Verify temporary MFA token
+      const payload = this.jwtService.verify<{
+        sub: string;
+        email: string;
+        role: string;
+        type: string;
+      }>(mfaToken, {
+        secret: authService.getJwtSecret(),
+      });
+
+      if (payload.type !== 'mfa_required') {
+        throw new UnauthorizedException('Invalid token type');
+      }
+
+      const user = await this.userRepository.findOne({
+        where: { id: payload.sub },
+      });
+
+      if (!user) {
+        throw new UnauthorizedException('User not found');
+      }
+
+      // Generate final tokens
+      const { accessToken, refreshToken } = authService.generateTokens(
+        user.id,
+        user.email,
+        user.role,
+      );
+
+      await authService.updateRefreshToken(user.id, refreshToken);
+
+      this.logger.log(`MFA login completed for user: ${user.id}`);
+
+      return {
+        user: authService.sanitizeUser(user),
+        accessToken,
+        refreshToken,
+      };
+    } catch (error: unknown) {
+      const message =
+        error instanceof Error ? error.message : 'Invalid or expired MFA token';
+      this.logger.error(`MFA login failed: ${message}`);
+      throw new UnauthorizedException('Invalid or expired MFA token');
+    }
   }
 }
